# app/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@ogc.org
    name: Open Geospatial Consortium
  description: Example API Definition for OGC API - Processes
  license:
    name: OGC License
    url: http://www.opengeospatial.org/legal/
  title: OGC API - Processes
  version: "0.1"
servers:
- description: "Example Server 1 (GNOSIS, supporting Part 1: Core /Sync and Part 3:\
    \ Workflows)"
  url: https://maps.gnosis.earth/ogcapi
paths:
...
  /processes:
    get:
      description: |
        The list of processes contains a summary of each process the OGC API - Processes offers, including the link to a more detailed description of the process.

        For more information, see [Section 7.7]https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_list).
      operationId: getProcesses
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/processList'
          description: Information about the available processes
      summary: retrieve the list of available processes
      tags:
      - Processes
    post:
      description: |
        Deploys a process.

        For more information, see [Section 6.3](http://docs.ogc.org/DRAFTS/20-044.html#_87a6983e-d060-458c-95ab-27e232e64822).
      operationId: deploy
      parameters:
      - description: Point to the workflow identifier for deploying a CWL containing
          multiple workflow definitions
        in: query
        name: w
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/ogcapppkg+json:
            schema:
              $ref: '#/components/schemas/ogcapppkg'
          application/cwl:
            schema:
              $ref: '#/components/schemas/cwl'
          application/cwl+json:
            schema:
              $ref: '#/components/schemas/cwl'
          application/cwl+yaml:
            schema:
              $ref: '#/components/schemas/cwl'
        description: An OGC Application Package used to deploy a new process.
        required: true
      responses:
        "201": {}
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: the processes is not mutable
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: the processes being added is already deployed (i.e. duplicate)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: deploy a process.
      tags:
      - DRU
  /processes/{processId}:
    delete:
      description: |
        Undeploys a process.

        For more information, see [Section 6.5](http://docs.ogc.org/DRAFTS/20-044.html#_16391f9e-538f-4a84-9710-72a6bab82842).
      operationId: undeploy
      parameters:
      - in: path
        name: processId
        required: true
        schema:
          $ref: '#/components/schemas/processes-list'
      responses:
        "204":
          description: successful operation (no response body)
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: the processes is not mutable
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: undeploy a process.
      tags:
      - DRU
    get:
      description: |
        The process description contains information about inputs and outputs and a link to the execution-endpoint for the process. The Core does not mandate the use of a specific process description to specify the interface of a process. That said, the Core requirements class makes the following recommendation:

        Implementations SHOULD consider supporting the OGC process description.

        For more information, see [Section 7.8](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_description).
      operationId: getProcessDescription
      parameters:
      - in: path
        name: processId
        required: true
        schema:
          $ref: '#/components/schemas/processes-list'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/process'
          description: A process description.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
      summary: retrieve a process description
      tags:
      - Processes
    put:
      description: |
        Replaces a process.

        For more information, see [Section 6.4](http://docs.ogc.org/DRAFTS/20-044.html#_18582f42-ebc6-4284-9333-c089068f62b6).
      operationId: replace
      parameters:
      - in: path
        name: processId
        required: true
        schema:
          $ref: '#/components/schemas/processes-list'
      requestBody:
        content:
          application/ogcapppkg+json:
            schema:
              $ref: '#/components/schemas/ogcapppkg'
          application/cwl:
            schema:
              $ref: '#/components/schemas/cwl'
          application/cwl+json:
            schema:
              $ref: '#/components/schemas/cwl'
          application/cwl+yaml:
            schema:
              $ref: '#/components/schemas/cwl'
        description: An OGC Application Package used to deploy a new process.
        required: true
      responses:
        "204":
          description: successful operation (no response body)
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: the processes is not mutable
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: "The requested resource does not exist on the server. For example,\
            \ a path parameter had an incorrect value."
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: the processes being added is already deployed (i.e. duplicate)
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
          description: A server error occurred.
      summary: replace a process.
      tags:
      - DRU
...
components:
...
  schemas:
...
    link:
      example:
        hreflang: en
        rel: service
        href: href
        type: application/json
        title: title
      properties:
        href:
          title: href
          type: string
        rel:
          example: service
          title: rel
          type: string
        type:
          example: application/json
          title: type
          type: string
        hreflang:
          example: en
          title: hreflang
          type: string
        title:
          title: title
          type: string
      required:
      - href
      title: link
      type: object
...
    exception:
      additionalProperties: true
      description: JSON schema for exceptions based on RFC 7807
      example:
        instance: instance
        detail: detail
        type: type
        title: title
        status: 0
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      required:
      - type
      title: Exception Schema
      type: object
...
    processes-list:
      enum:
      - RenderMap
      - ElevationContours
      - OSMERE
      title: processes-list
      type: string
    processSummary:
      allOf:
      - $ref: '#/components/schemas/descriptionType'
      - properties:
          id:
            type: string
          version:
            type: string
          jobControlOptions:
            items:
              $ref: '#/components/schemas/jobControlOptions'
            type: array
          links:
            items:
              $ref: '#/components/schemas/link'
            type: array
        required:
        - id
        - version
        type: object
      example:
        metadata:
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        keywords:
        - keywords
        - keywords
        description: description
        links:
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        id: id
        title: title
        version: version
        jobControlOptions:
        - sync-execute
        - sync-execute
      title: processSummary
    process:
      allOf:
      - $ref: '#/components/schemas/processSummary'
      - properties:
          inputs:
            additionalProperties:
              $ref: '#/components/schemas/inputDescription'
            type: object
          outputs:
            additionalProperties:
              $ref: '#/components/schemas/outputDescription'
            type: object
        type: object
      example:
        outputs:
          key:
            schema: null
            metadata:
            - null
            - null
            keywords:
            - keywords
            - keywords
            description: description
            title: title
        metadata:
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          role: role
          rel: service
          href: href
          type: application/json
          title: title
        keywords:
        - keywords
        - keywords
        inputs:
          key:
            schema:
              $ref: $ref
            metadata:
            - null
            - null
            keywords:
            - keywords
            - keywords
            minOccurs: 0
            valuePassing:
            - byValue
            - byValue
            description: description
            maxOccurs: 6
            title: title
        description: description
        links:
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        id: id
        title: title
        version: version
        jobControlOptions:
        - sync-execute
        - sync-execute
      title: process
    processList:
      example:
        processes:
        - metadata:
          - hreflang: en
            role: role
            rel: service
            href: href
            type: application/json
            title: title
          - hreflang: en
            role: role
            rel: service
            href: href
            type: application/json
            title: title
          keywords:
          - keywords
          - keywords
          description: description
          links:
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          id: id
          title: title
          version: version
          jobControlOptions:
          - sync-execute
          - sync-execute
        - metadata:
          - hreflang: en
            role: role
            rel: service
            href: href
            type: application/json
            title: title
          - hreflang: en
            role: role
            rel: service
            href: href
            type: application/json
            title: title
          keywords:
          - keywords
          - keywords
          description: description
          links:
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          - hreflang: en
            rel: service
            href: href
            type: application/json
            title: title
          id: id
          title: title
          version: version
          jobControlOptions:
          - sync-execute
          - sync-execute
        links:
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
        - hreflang: en
          rel: service
          href: href
          type: application/json
          title: title
      properties:
        processes:
          items:
            $ref: '#/components/schemas/processSummary'
          title: processes
          type: array
        links:
          items:
            $ref: '#/components/schemas/link'
          title: links
          type: array
      required:
      - links
      - processes
      title: processList
      type: object
...
    descriptionType:
      properties:
        title:
          title: title
          type: string
        description:
          title: description
          type: string
        keywords:
          items:
            type: string
          title: keywords
          type: array
        metadata:
          items:
            $ref: '#/components/schemas/metadata'
          title: metadata
          type: array
      title: descriptionType
      type: object
    binaryInputValue:
      format: byte
      type: string
...
    format:
      properties:
        mediaType:
          title: mediaType
          type: string
        encoding:
          title: encoding
          type: string
        schema:
          $ref: '#/components/schemas/format_schema'
      title: format
      type: object
    inputDescription:
      allOf:
      - $ref: '#/components/schemas/descriptionType'
      - properties:
          schema:
            $ref: '#/components/schemas/schema'
          minOccurs:
            default: 1
            type: integer
          maxOccurs:
            $ref: '#/components/schemas/inputDescription_allOf_maxOccurs'
          valuePassing:
            default:
            - byValue
            - byReference
            items:
              enum:
              - byValue
              - byReference
              type: string
            type: array
        required:
        - schema
        type: object
      example:
        schema:
          $ref: $ref
        metadata:
        - null
        - null
        keywords:
        - keywords
        - keywords
        minOccurs: 0
        valuePassing:
        - byValue
        - byValue
        description: description
        maxOccurs: 6
        title: title
      title: inputDescription
    inputValue:
      anyOf:
      - $ref: '#/components/schemas/inputValueNoObject'
      - type: object
      title: inputValue
    inputValueNoObject:
      oneOf:
      - type: string
      - type: number
      - type: integer
      - type: boolean
      - items:
          type: object
        type: array
      - $ref: '#/components/schemas/binaryInputValue'
      - $ref: '#/components/schemas/bbox'
      title: inputValueNoObject
    jobControlOptions:
      enum:
      - sync-execute
      - async-execute
      - dismiss
      title: jobControlOptions
      type: string
    metadata:
      oneOf:
      - $ref: '#/components/schemas/metadata_oneOf'
      - $ref: '#/components/schemas/metadata_oneOf_1'
      title: metadata
...
    outputDescription:
      allOf:
      - $ref: '#/components/schemas/descriptionType'
      - properties:
          schema:
            $ref: '#/components/schemas/schema'
        required:
        - schema
        type: object
      example:
        schema: null
        metadata:
        - null
        - null
        keywords:
        - keywords
        - keywords
        description: description
        title: title
      title: outputDescription
    qualifiedInputValue:
      allOf:
      - $ref: '#/components/schemas/format'
      - properties:
          value:
            $ref: '#/components/schemas/inputValue'
        required:
        - value
        type: object
      title: qualifiedInputValue
    reference:
      additionalProperties: false
      example:
        $ref: $ref
      properties:
        $ref:
          format: uri-reference
          title: $ref
          type: string
      required:
      - $ref
      title: reference
      type: object
...
    schema:
      oneOf:
      - $ref: '#/components/schemas/reference'
      - $ref: '#/components/schemas/schema_oneOf'
      title: schema
...
    ogcapppkg:
      properties:
        processDescription:
          $ref: '#/components/schemas/process'
        executionUnit:
          $ref: '#/components/schemas/ogcapppkg_executionUnit'
      required:
      - executionUnit
      title: ogcapppkg
      type: object
...
    cwl:
      type: object
...
    bbox:
      properties:
        bbox:
          items:
            type: number
          oneOf:
          - maxItems: 4
            minItems: 4
            type: object
          - maxItems: 6
            minItems: 6
            type: object
          title: bbox
          type: array
        crs:
          $ref: '#/components/schemas/bbox-def-crs'
      required:
      - bbox
      title: bbox
      type: object
    schema_1:
      oneOf:
      - $ref: '#/components/schemas/reference'
      - $ref: '#/components/schemas/schema_oneOf'
      title: schema_1
    executionUnit:
      additionalProperties: true
      description: Resource containing an executable or runtime information for executing
        the process.
      example:
        type: docker
        image: mydocker/ndvi:latest
        deployment: local
        config:
          cpuMin: 2
          cpuMax: 5
          memoryMin: 1
          memoryMax: 3
      properties:
        type:
          description: Type of execution unit.
          enum:
          - docker
          - oci
          type: string
        image:
          description: Container image reference for the execution unit.
          type: string
        deployment:
          description: Deployment information for the execution unit.
          enum:
          - local
          - remote
          - hpc
          - cloud
          type: string
        config:
          $ref: '#/components/schemas/executionUnit_config'
      required:
      - image
      - type
      title: executionUnit
      type: object
    ogcapppkg-array:
      items:
        $ref: '#/components/schemas/ogcapppkg_array_inner'
      type: array
...
    bbox-def-crs:
      anyOf:
      - default: http://www.opengis.net/def/crs/OGC/1.3/CRS84
        enum:
        - http://www.opengis.net/def/crs/OGC/1.3/CRS84
        - http://www.opengis.net/def/crs/OGC/0/CRS84h
        format: uri
        type: string
      - default: http://www.opengis.net/def/crs/OGC/1.3/CRS84
        format: uri
        type: string
      title: bbox-def-crs
...
    format_schema:
      oneOf:
      - format: url
        type: string
      - type: object
      title: format_schema
    inputDescription_allOf_maxOccurs:
      oneOf:
      - default: 1
        type: integer
      - enum:
        - unbounded
        type: string
      title: inputDescription_allOf_maxOccurs
    metadata_oneOf:
      allOf:
      - $ref: '#/components/schemas/link'
      - properties:
          role:
            type: string
        type: object
      example:
        hreflang: en
        role: role
        rel: service
        href: href
        type: application/json
        title: title
      title: metadata_oneOf
    metadata_oneOf_1_value:
      oneOf:
      - type: string
      - type: object
      title: metadata_oneOf_1_value
    metadata_oneOf_1:
      properties:
        role:
          title: role
          type: string
        title:
          title: title
          type: string
        lang:
          title: lang
          type: string
        value:
          $ref: '#/components/schemas/metadata_oneOf_1_value'
      title: metadata_oneOf_1
      type: object
    schema_oneOf_additionalProperties:
      default: true
      oneOf:
      - $ref: '#/components/schemas/schema_1'
      - type: boolean
      title: schema_oneOf_additionalProperties
    schema_oneOf:
      additionalProperties: false
      properties:
        title:
          title: title
          type: string
        multipleOf:
          exclusiveMinimum: true
          minimum: 0
          title: multipleOf
          type: number
        maximum:
          title: maximum
          type: number
        exclusiveMaximum:
          default: false
          title: exclusiveMaximum
          type: boolean
        minimum:
          title: minimum
          type: number
        exclusiveMinimum:
          default: false
          title: exclusiveMinimum
          type: boolean
        maxLength:
          minimum: 0
          title: maxLength
          type: integer
        minLength:
          default: 0
          minimum: 0
          title: minLength
          type: integer
        pattern:
          format: regex
          title: pattern
          type: string
        maxItems:
          minimum: 0
          title: maxItems
          type: integer
        minItems:
          default: 0
          minimum: 0
          title: minItems
          type: integer
        uniqueItems:
          default: false
          title: uniqueItems
          type: boolean
        maxProperties:
          minimum: 0
          title: maxProperties
          type: integer
        minProperties:
          default: 0
          minimum: 0
          title: minProperties
          type: integer
        required:
          items:
            type: string
          minItems: 1
          title: required
          type: array
          uniqueItems: true
        enum:
          items:
            type: object
          minItems: 1
          title: enum
          type: array
          uniqueItems: false
        type:
          enum:
          - array
          - boolean
          - integer
          - number
          - object
          - string
          title: type
          type: string
        not:
          $ref: '#/components/schemas/schema_1'
        allOf:
          items:
            $ref: '#/components/schemas/schema_1'
          title: allOf
          type: array
        oneOf:
          items:
            $ref: '#/components/schemas/schema_1'
          title: oneOf
          type: array
        anyOf:
          items:
            $ref: '#/components/schemas/schema_1'
          title: anyOf
          type: array
        items:
          $ref: '#/components/schemas/schema_1'
        properties:
          additionalProperties:
            $ref: '#/components/schemas/schema_1'
          properties: {}
          title: properties
          type: object
        additionalProperties:
          $ref: '#/components/schemas/schema_oneOf_additionalProperties'
        description:
          title: description
          type: string
        format:
          title: format
          type: string
        default:
          title: default
          type: object
        nullable:
          default: false
          title: nullable
          type: boolean
        readOnly:
          default: false
          title: readOnly
          type: boolean
        writeOnly:
          default: false
          title: writeOnly
          type: boolean
        example:
          title: example
          type: object
        deprecated:
          default: false
          title: deprecated
          type: boolean
        contentMediaType:
          title: contentMediaType
          type: string
        contentEncoding:
          title: contentEncoding
          type: string
        contentSchema:
          title: contentSchema
          type: string
      title: schema_oneOf
      type: object
    ogcapppkg_executionUnit:
      oneOf:
      - $ref: '#/components/schemas/executionUnit'
      - $ref: '#/components/schemas/link'
      - $ref: '#/components/schemas/qualifiedInputValue'
      - $ref: '#/components/schemas/ogcapppkg-array'
      title: ogcapppkg_executionUnit
...
    executionUnit_config:
      additionalProperties: true
      description: Hardware requirements and configuration properties for executing
        the process.
      properties:
        cpuMin:
          description: Minimum number of CPUs required to run the process (unit is
            CPU core).
          minimum: 1
          type: number
        cpuMax:
          description: Maximum number of CPU dedicated to the process (unit is CPU
            core)
          type: number
        memoryMin:
          description: Minimum RAM memory required to run the application (unit is
            GB)
          type: number
        memoryMax:
          description: Maximum RAM memory dedicated to the application (unit is GB)
          type: number
        storageTempMin:
          description: Minimum required temporary storage size (unit is GB)
          type: number
        storageOutputsMin:
          description: Minimum required output storage size (unit is GB)
          type: number
        jobTimeout:
          description: Timeout delay for a job execution (in seconds)
          type: number
      title: executionUnit_config
      type: object
    ogcapppkg_array_inner:
      oneOf:
      - $ref: '#/components/schemas/executionUnit'
      - $ref: '#/components/schemas/link'
      - $ref: '#/components/schemas/qualifiedInputValue'
      title: ogcapppkg_array_inner
...